feature/88 db week 엔티티 데이터 컬럼 관련 오류 메서드 수정

+++ src/main/java/homeat/backend/domain/homeatreport/dto/ReportMonthlyAnalyzeResponseDTO.java
@@ -13,8 +13,8 @@ public class ReportMonthlyAnalyzeResponseDTO {
 
     private Long month_jipbap_price;
     private Long month_out_price;
-    private double jipbap_ratio;
-    private double out_ratio;
+    private int jipbap_ratio;
+    private int out_ratio;
     private String save_percent;
 
 }
